requirements.txt
python-dotenv
openai
faiss
flask